refactor: Person 위치 변경
