+++ src/main/java/persistence/sql/Person.java
@@ -1,10 +1,21 @@
-package persistence.domain;
+package persistence.sql;
 
 import jakarta.persistence.*;
 
 @Table(name = "users")
 @Entity
 public class Person {
+
+    public Person() {
+    }
+
+    public Person(Long id, String name, Integer age, String email, Integer index) {
+        this.id = id;
+        this.name = name;
+        this.age = age;
+        this.email = email;
+        this.index = index;
+    }
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
