+++ src/main/java/persistence/sql/entity/Entity.java
@@ -1,0 +1,2 @@
+package persistence.sql.entity;public class Entity {
+}
