+++ src/test/java/persistence/sql/ddl/DropQueryBuilderTest.java
@@ -1,0 +1,7 @@
+package persistence.sql.ddl;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class DropQueryBuilderTest {
+
+}
