+++ src/test/java/persistence/sql/ddl/TableTest.java
@@ -5,7 +5,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import persistence.domain.Person;
+import persistence.sql.Person;
 import persistence.sql.domain.Table;
 
 import java.lang.reflect.Field;
